requirements.txt
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pybind11==2.13.6
pyparsing==3.2.1
pyroomacoustics==0.8.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
wheel==0.45.1
